Hacking in a few IP Options.
Changed congestionLevel from int to double.
Added tags for IP Option RouteRecord and Strict source routing.
Trying to hack-in RouteRecord on all TCP-Data and Strict Source routing
on Acks. (Need to propagate the RouteRecord to TCP and copy it to the
Ack as Strict source).

At least it does not crash ...

Adding requested IPoptions tags to Options in IP Header
Adding RouteRecord reflector to TCP receiver (copy Route record from
Data segment to Strict Source Routing in Ack)

Dmpr: Fix: Incorrect access to interface protocol data

This is connected to refactoring the code to template based
protocol interface data.

+++ src/inet/routing/dmpr/DmprInterfaceData.cc
@@ -24,13 +24,13 @@ DmprInterfaceData::DmprInterfaceData()
 
 }
 
-int DmprInterfaceData::getCongestionLevel() const
+double DmprInterfaceData::getCongestionLevel() const
 {
   return congestionLevel;
 }
 
 
-void DmprInterfaceData::setCongestionLevel(int congestionLevel)
+void DmprInterfaceData::setCongestionLevel(double congestionLevel)
 {
   this->congestionLevel = congestionLevel;
 }
